refactor user modules
Had issues with conditional imports of modules per-user. To resolve this, refactored most of the modules to be included in "default" and toggled via a switch.

In retrospect, the issues were most likely due to a circular reference when using users.users as the condition for importing. It may possible to go back to seperate modules and conditional imports at a later date.